一時ファイルの保存時に内容を保持して、ディスクIOを減らす

+++ src/save.pl
@@ -1,20 +1,22 @@
 
 my @saved_files = ();
+my %saved_files_content = ();
+my $save_file_dryrun = 1;
 
 sub save_file {
-    my ($content, $ext) = @_;
+    my ($content, $ext, $keep) = @_;
 
     my ($fh, $filename) = tempfile;
     print $fh encode('utf-8', $content);
     close $fh;
 
-    save_file_from_tempfile($filename, $ext);
+    save_file_from_tempfile($filename, $ext, $keep, $content);
 }
 
 sub save_file_from_tempfile {
-    my ($filename, $ext) = @_;
+    my ($filename, $ext, $keep, $content) = @_;
 
-    my $key=`sha1sum $filename`;
+    my $key=`sha1sum $filename`; # TODO ファイルに保存しなくてもsha1をとれるようにすべき
     die unless ($key =~ /\A([^\s]+)/);
     $key = $1;
 
@@ -31,7 +33,17 @@ sub save_file_from_tempfile {
         $target_name = "$prefix$key";
     }
 
-    if ( -e $target_path) {
+    unless (defined($content)) {
+        die; # TODO
+    }
+
+    if ($keep) {
+        $saved_files_content{$target_name} = $content;
+    }
+
+    if ($save_file_dryrun) {
+        `rm $filename`;
+    } elsif ( -e $target_path) {
         `rm $filename`;
     } else {
         `mv $filename $target_path`;
@@ -42,25 +54,27 @@ sub save_file_from_tempfile {
     $target_name;
 }
 
+sub get_saved_file {
+    my ($name) = @_;
+    $saved_files_content{$name};
+}
+
 sub get_comments_about_saved_files {
     my ($lang) = @_;
     my $roonda_tmp_path = $ENV{$ENV_TMP_PATH};
     my $result = "";
-    foreach my $file_name (@saved_files) {
-        open(IN, '<', "$roonda_tmp_path/$file_name") or die "Cannot open";
-        my @lines = <IN>;
-        close IN;
+    foreach my $file_name (keys %saved_files_content) {
+        my $content = $saved_files_content{$file_name};
         $result = $result . "\n";
-        my $source = join('', @lines);
         my $source_comment;
         if ($lang eq $LANG_SH) {
             my $splitter = "#################################################";
             my $header = "$splitter\n# $file_name:\n$splitter\n";
-            $source_comment = $header . escape_sh_multiline_comment($source);
+            $source_comment = $header . escape_sh_multiline_comment($content);
         } elsif ($lang eq $LANG_PERL) {
             my $splitter = "#################################################";
             my $header = "$splitter\n# $file_name:\n$splitter\n";
-            $source_comment = $header . escape_perl_multiline_comment($source);
+            $source_comment = $header . escape_perl_multiline_comment($content);
         } else {
             die;
         }
